new: merge phosphor motion blur into PolyBlur - misc: reorganize structure of blurs - new: allow force disabling fast render - fix: fix alpha not being enabled

+++ src/main/resources/assets/polyblur/shaders/monkeyblur.frag
@@ -11,7 +11,7 @@ uniform mat4 previousModelView;
 uniform vec3 cameraPosition;
 uniform vec3 previousCameraPosition;
 
-uniform int strength;
+uniform float strength;
 
 varying vec4 texcoord;
